test(e2e): consolidate exec e2e into sandbox.e2e.test.ts; tidy example comments
All sandbox e2e (exec/durable, fork, template) now live in one file. Remove
em dashes and verbose comments from the examples for a concise, professional tone.
