+++ examples/sandboxes/exec.ts
@@ -1,9 +1,8 @@
 import { Sandbox } from "../../src/index.ts";
 import { runExample, sleep } from "./helpers.ts";
 
-// Durable exec: a command runs on the sandbox independently of any client.
-// Start one, detach, then reattach from a fresh connection (even another
-// process) by its sessionName and stream it through to completion.
+// Durable exec: a command runs independently of the client. Start one, detach,
+// then reattach by sessionName from a fresh connection and stream to completion.
 await runExample(async () => {
   await using sandbox = await Sandbox.create();
 
@@ -11,14 +10,14 @@ await runExample(async () => {
     "for i in $(seq 1 20); do echo line-$i; sleep 0.5; done",
   );
   const sessionName = await handle.sessionName;
-  await handle.detach(); // stop watching; the command keeps running (kill() ends it)
-  console.log(`detached — ${sessionName} keeps running\n`);
+  await handle.detach(); // stop streaming without ending the command
+  console.log(`detached, ${sessionName} still running\n`);
 
-  // The command runs on the sandbox while we're away — reconnect after a beat.
+  // The command keeps running while disconnected; reconnect after a delay.
   await sleep(1_000);
 
   // Reattach by sessionName: replays the log so far, then follows live.
-  // (pass resumeFromLastRead: true for only new output instead of a full replay.)
+  // Pass resumeFromLastRead: true to receive only new output.
   const reattached = await Sandbox.connect(sandbox.id);
   const result = await reattached.exec(
     { sessionName },
